requirements.txt
anyio==3.6.2
aiofiles>=23.2.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.3
fastapi==0.115.6
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httptools==0.5.0
httpx==0.28.1
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==1.25.2
pandas==2.1.0
psycopg2-binary==2.9.10
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0
requests==2.32.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
Werkzeug==3.1.3
python-jose==3.4.0